requirements.txt
streamlit==1.36.0
pandas
numpy
scikit-learn
joblib
xgboost==2.0.3
python-docx
